# openapi.json
{"openapi":"3.1.0","info":{"title":"MultiAgentPlatform","description":"High-Performance Multi-Agent Platform API","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["System"],"summary":"Health Check","description":"Basic health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/api/v1/tasks":{"post":{"tags":["Tasks"],"summary":"Submit task","operationId":"create_task_api_v1_tasks_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskSubmittedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/{task_id}":{"get":{"tags":["Tasks"],"summary":"Get Task Status","operationId":"get_task_status_api_v1_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contexts/{context_id}":{"get":{"tags":["Context Management"],"summary":"Get context by ID","operationId":"get_context_by_id_api_v1_contexts__context_id__get","parameters":[{"name":"context_id","in":"path","required":true,"schema":{"type":"string","description":"Unique context ID","title":"Context Id"},"description":"Unique context ID"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Context Management"],"summary":"Create or update context","operationId":"create_or_update_context_api_v1_contexts__context_id__post","parameters":[{"name":"context_id","in":"path","required":true,"schema":{"type":"string","description":"Context ID","title":"Context Id"},"description":"Context ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"description":"Arbitrary JSON payload","title":"Context Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextOperationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agents":{"get":{"tags":["Agent Management"],"summary":"List Registered Agents","description":"Retrieves a list of all registered agent configurations.","operationId":"list_registered_agents_api_v1_agents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AgentInfo"},"type":"array","title":"Response List Registered Agents Api V1 Agents Get"}}}}}},"post":{"tags":["Agent Management"],"summary":"Register New Agent","description":"Register a new agent configuration from JSON data.","operationId":"register_agent_api_v1_agents_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentConfig"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agents/{agent_name}":{"get":{"tags":["Agent Management"],"summary":"Get Agent Configuration","description":"Retrieves the detailed configuration for a specific agent by its name.","operationId":"get_agent_configuration_api_v1_agents__agent_name__get","parameters":[{"name":"agent_name","in":"path","required":true,"schema":{"type":"string","title":"Agent Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentDetailResponse"}}}},"404":{"description":"Agent configuration not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tools":{"get":{"tags":["Tool Management & Execution"],"summary":"List Available Tools","description":"Retrieves a list of all tools registered in the system with basic information.","operationId":"list_available_tools_api_v1_tools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ToolInfo"},"type":"array","title":"Response List Available Tools Api V1 Tools Get"}}}}}}},"/api/v1/tools/{tool_name}":{"get":{"tags":["Tool Management & Execution"],"summary":"Get Tool Details","operationId":"get_tool_details_api_v1_tools__tool_name__get","parameters":[{"name":"tool_name","in":"path","required":true,"schema":{"type":"string","title":"Tool Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tools/{tool_name}/execute":{"post":{"tags":["Tool Management & Execution"],"summary":"Execute a tool immediately","operationId":"execute_tool_api_v1_tools__tool_name__execute_post","parameters":[{"name":"tool_name","in":"path","required":true,"schema":{"type":"string","title":"Tool Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolExecutionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolExecutionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/config":{"get":{"tags":["System Configuration"],"summary":"Get System Configuration","description":"Retrieves the current system configuration settings, excluding sensitive information like API keys and passwords.","operationId":"get_system_configuration_api_v1_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/metrics":{"get":{"tags":["Metrics"],"summary":"Get Metrics","operationId":"get_metrics_api_v1_metrics_get","parameters":[{"name":"accept","in":"header","required":false,"schema":{"type":"string","default":"text/plain","title":"Accept"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AgentCapability":{"type":"string","enum":["planning","execution","reasoning","tool_use","code_generation","information_retrieval","creative_writing","conversation","summarization","text_analysis"],"title":"AgentCapability"},"AgentConfig":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description","default":""},"version":{"type":"string","title":"Version","default":"1.0.0"},"agent_type":{"type":"string","title":"Agent Type"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"capabilities":{"items":{"$ref":"#/components/schemas/AgentCapability"},"type":"array","uniqueItems":true,"title":"Capabilities"},"parameters":{"additionalProperties":true,"type":"object","title":"Parameters"},"max_retries":{"type":"integer","title":"Max Retries","default":3},"timeout":{"type":"number","title":"Timeout","default":30.0},"allowed_tools":{"items":{"type":"string"},"type":"array","title":"Allowed Tools"},"memory_keys":{"items":{"type":"string"},"type":"array","title":"Memory Keys"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"},"mcp_enabled":{"type":"boolean","title":"Mcp Enabled","description":"Whether this agent instance uses the Model Context Protocol.","default":false},"mcp_context_types":{"items":{"type":"string"},"type":"array","title":"Mcp Context Types","description":"List of primary MCP context types this agent interacts with."}},"additionalProperties":true,"type":"object","required":["name","agent_type"],"title":"AgentConfig"},"AgentDetailResponse":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description","default":""},"version":{"type":"string","title":"Version","default":"1.0.0"},"agent_type":{"type":"string","title":"Agent Type"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"capabilities":{"items":{"$ref":"#/components/schemas/AgentCapability"},"type":"array","uniqueItems":true,"title":"Capabilities"},"parameters":{"additionalProperties":true,"type":"object","title":"Parameters"},"max_retries":{"type":"integer","title":"Max Retries","default":3},"timeout":{"type":"number","title":"Timeout","default":30.0},"allowed_tools":{"items":{"type":"string"},"type":"array","title":"Allowed Tools"},"memory_keys":{"items":{"type":"string"},"type":"array","title":"Memory Keys"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"},"mcp_enabled":{"type":"boolean","title":"Mcp Enabled","description":"Whether this agent instance uses the Model Context Protocol.","default":false},"mcp_context_types":{"items":{"type":"string"},"type":"array","title":"Mcp Context Types","description":"List of primary MCP context types this agent interacts with."}},"additionalProperties":true,"type":"object","required":["name","agent_type"],"title":"AgentDetailResponse"},"AgentInfo":{"properties":{"name":{"type":"string","title":"Name"},"agent_type":{"type":"string","title":"Agent Type"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"version":{"type":"string","title":"Version"}},"type":"object","required":["name","agent_type","version"],"title":"AgentInfo"},"ContextOperationResponse":{"properties":{"context_id":{"type":"string","title":"Context Id"},"status":{"type":"string","enum":["created","updated"],"title":"Status"},"message":{"type":"string","title":"Message"}},"type":"object","required":["context_id","status","message"],"title":"ContextOperationResponse"},"ContextResponse":{"properties":{"context_id":{"type":"string","title":"Context Id"},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["context_id","data"],"title":"ContextResponse"},"CreateTaskRequest":{"properties":{"goal":{"type":"string","title":"Goal"},"task_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Type"},"input_data":{"additionalProperties":true,"type":"object","title":"Input Data"},"priority":{"anyOf":[{"$ref":"#/components/schemas/TaskPriority"},{"type":"integer"}],"title":"Priority","default":"NORMAL"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","required":["goal"],"title":"CreateTaskRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","title":"Status","default":"ok"}},"type":"object","title":"HealthCheckResponse"},"TaskPriority":{"type":"string","enum":["LOW","NORMAL","HIGH","CRITICAL"],"title":"TaskPriority"},"TaskState":{"type":"string","enum":["PENDING","RUNNING","COMPLETED","FAILED"],"title":"TaskState"},"TaskStatusResponse":{"properties":{"id":{"type":"string","title":"Id"},"state":{"$ref":"#/components/schemas/TaskState"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result"},"error":{"anyOf":[{"type":"string"},{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Error"}},"type":"object","required":["id","state"],"title":"TaskStatusResponse","description":"Response model for task status"},"TaskSubmittedResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","const":"submitted","title":"Status"}},"type":"object","required":["task_id","status"],"title":"TaskSubmittedResponse"},"ToolExecutionRequest":{"properties":{"args":{"additionalProperties":true,"type":"object","title":"Args","description":"Arguments to pass to the tool, matching its args_schema."}},"type":"object","title":"ToolExecutionRequest"},"ToolExecutionResponse":{"properties":{"status":{"type":"string","title":"Status","description":"'success' or 'error'"},"tool_name":{"type":"string","title":"Tool Name"},"execution_time":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Execution Time","description":"Execution time in seconds (if available)"},"result":{"anyOf":[{},{"type":"null"}],"title":"Result","description":"Result of the tool execution (if successful)"},"error":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Error","description":"Error details (if execution failed)"}},"type":"object","required":["status","tool_name"],"title":"ToolExecutionResponse"},"ToolInfo":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"args_schema_summary":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Args Schema Summary"}},"type":"object","required":["name","description"],"title":"ToolInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}